+++ docs/getting-started/changelog.md
@@ -1,4 +1,10 @@
 # Changelog
+## Version 1.5.2
+### Released 11 February 2021
+- Preference Variables
+	- xprefOrientationChangeTransition - When a screen orientation occurs, smoothly transition from one state to the other
+	- xprefDocumentBackgroundColor - Change the color of the HTML document background
+	
 ## Version 1.5.1
 ### Released 1 December 2020
 - Command Variables